requirements.txt
azure-eventhub==5.15.1
azure-eventhub-checkpointstoreblob==1.2.0
azure-identity==1.25.3
azure-storage-file-datalake==12.23.0
confluent-kafka==2.14.0
pyarrow==24.0.0
prometheus-client==0.25.0
pytest==9.0.3
